requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.8
django-session-timeout==0.1.0
idna==3.4
Pillow==10.0.1
python-decouple==3.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7
